+++ src/main/java/com/epam/reportportal/service/ReportPortal.java
@@ -91,7 +91,7 @@ public class ReportPortal {
 			@Nullable LaunchIdLock launchIdLock) {
 		this.rpClient = rpClient;
 		this.executor = executor;
-		this.parameters = parameters;
+		this.parameters = Objects.requireNonNull(parameters);
 		this.launchIdLock = launchIdLock;
 	}
 
@@ -101,7 +101,8 @@ public class ReportPortal {
 	 * @param rq Request Data
 	 * @return Launch
 	 */
-	public Launch newLaunch(StartLaunchRQ rq) {
+	@Nonnull
+	public Launch newLaunch(@Nonnull StartLaunchRQ rq) {
 		if (BooleanUtils.isNotTrue(parameters.getEnable()) || rpClient == null) {
 			return Launch.NOOP_LAUNCH;
 		}
@@ -144,13 +145,15 @@ public class ReportPortal {
 	 * @param launchUuid Launch to be used
 	 * @return This instance for chaining
 	 */
-	public Launch withLaunch(Maybe<String> launchUuid) {
+	@Nonnull
+	public Launch withLaunch(@Nonnull Maybe<String> launchUuid) {
 		return ofNullable(rpClient).map(c -> (Launch) new LaunchImpl(c, parameters, launchUuid, executor)).orElse(Launch.NOOP_LAUNCH);
 	}
 
 	/**
 	 * @return Configuration parameters
 	 */
+	@Nonnull
 	public ListenerParameters getParameters() {
 		return parameters;
 	}
@@ -158,6 +161,7 @@ public class ReportPortal {
 	/**
 	 * @return ReportPortal client
 	 */
+	@Nullable
 	public ReportPortalClient getClient() {
 		return this.rpClient;
 	}
@@ -167,6 +171,7 @@ public class ReportPortal {
 	 *
 	 * @return builder for {@link ReportPortal}
 	 */
+	@Nonnull
 	public static Builder builder() {
 		return new Builder();
 	}
@@ -182,6 +187,7 @@ public class ReportPortal {
 	 * @param params {@link ListenerParameters}
 	 * @return builder for {@link ReportPortal}
 	 */
+	@Nonnull
 	public static ReportPortal create(ReportPortalClient client, ListenerParameters params) {
 		return create(client, params, buildExecutorService(params));
 	}
@@ -194,6 +200,7 @@ public class ReportPortal {
 	 * @param executor An executor service which will be used for internal request / response queue
 	 * @return builder for {@link ReportPortal}
 	 */
+	@Nonnull
 	public static ReportPortal create(@Nonnull final ReportPortalClient client, @Nonnull final ListenerParameters params,
 			@Nonnull final ExecutorService executor) {
 		return new ReportPortal(client, executor, params, getLaunchLock(params));